backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
